PD-40 Remove server-side TLPK support
