+++ server/Enums.js
@@ -4,11 +4,6 @@ module.exports = {
 
   DEPLOYMENT_INSTANCES_LIST_MAXIMUM_LENGTH: 300,
 
-  SourcePackageType: {
-    CodeDeployRevision: 'CodeDeployRevision',
-    DeploymentMap: 'DeploymentMap'
-  },
-
   DeploymentMode: {
     Overwrite: 'overwrite',
     BlueGreen: 'bg'
