--- server/test/modules/dm-packer/test-data/package-metadata-example.xml
@@ -1,36 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<entry xmlns="http://www.w3.org/2005/Atom" xmlns:d="http://schemas.microsoft.com/ado/2007/08/dataservices" xmlns:m="http://schemas.microsoft.com/ado/2007/08/dataservices/metadata" xml:base="http://view.artifacts.ttldev/artifactory/api/nuget/ttlpackages-internal-master/">
-  <id>http://view.artifacts.ttldev/artifactory/api/nuget/ttlpackages-internal-master/Packages(Id='Trainline.AgentAdmin.tlpk',Version='1.0.2')</id>
-  <title type="text">Trainline.AgentAdmin.tlpk</title>
-  <summary type="text">Trainline.AgentAdmin.tlpk</summary>
-  <updated>2016-04-15T18:03:08Z</updated>
-  <author>
-    <name>thetrainline.com</name>
-  </author>
-  <link rel="edit" title="V2FeedPackage" href="Packages(Id='Trainline.AgentAdmin.tlpk',Version='1.0.2')"/>
-  <category term="NuGetGallery.V2FeedPackage" scheme="http://schemas.microsoft.com/ado/2007/08/dataservices/scheme"/>
-  <content type="application/zip" src="http://view.artifacts.ttldev/artifactory/api/nuget/ttlpackages-internal-master/Download/Trainline.AgentAdmin.tlpk/1.0.2"/>
-  <m:properties xmlns:m="http://schemas.microsoft.com/ado/2007/08/dataservices/metadata" xmlns:d="http://schemas.microsoft.com/ado/2007/08/dataservices">
-    <d:lastUpdated>2016-04-15T18:03:08</d:lastUpdated>
-    <d:Version>1.0.2</d:Version>
-    <d:Copyright>Copyright 2016</d:Copyright>
-    <d:Created m:type="Edm.DateTime">2016-04-15T18:03:09</d:Created>
-    <d:Dependencies>TTL.AccountLinkingService.tlpk:1.0.0-alpha056|Trainline.AppInstaller:8.6.111</d:Dependencies>
-    <d:Description>Trainline.AgentAdmin.tlpk</d:Description>
-    <d:DownloadCount m:type="Edm.Int32">0</d:DownloadCount>
-    <d:IsLatestVersion m:type="Edm.Boolean">true</d:IsLatestVersion>
-    <d:IsAbsoluteLatestVersion m:type="Edm.Boolean">true</d:IsAbsoluteLatestVersion>
-    <d:IsPrerelease m:type="Edm.Boolean">false</d:IsPrerelease>
-    <d:Language m:null="true"/>
-    <d:Published m:type="Edm.DateTime">2016-04-15T18:03:09</d:Published>
-    <d:PackageHash>JEmvN5hmrleC5yXt68BExkEJVWB98IKVG3rCT+60Ar3uaEqfZXSbMgNCSnXzpQXwetq0ry3Ao9wbAmrtel39Eg==</d:PackageHash>
-    <d:PackageHashAlgorithm>SHA512</d:PackageHashAlgorithm>
-    <d:PackageSize m:type="Edm.Int64">4310708</d:PackageSize>
-    <d:RequireLicenseAcceptance m:type="Edm.Boolean">false</d:RequireLicenseAcceptance>
-    <d:Tags>Trainline.AgentAdmin.tlpk</d:Tags>
-    <d:Title m:null="true"/>
-    <d:VersionDownloadCount m:type="Edm.Int32">0</d:VersionDownloadCount>
-    <d:Authors>thetrainline.com</d:Authors>
-    <d:MinClientVersion m:null="true"/>
-  </m:properties>
-</entry>
